+++ app/src/main/java/com/maheswara660/packora/MainActivity.kt
@@ -274,9 +274,9 @@ fun incrementVersionString(v: String): String {
 fun Context.appVersion(): String {
     return try {
         val pInfo = packageManager.getPackageInfo(packageName, 0)
-        pInfo.versionName ?: "3.1.0"
+        pInfo.versionName ?: "3.1.1"
     } catch (e: Exception) {
-        "3.1.0"
+        "3.1.1"
     }
 }